backend/requirements.txt
# Core Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1

# LangChain & LangGraph
langchain==0.3.13
langchain-core==0.3.28
langchain-openai==0.2.14
langchain-mistralai==0.2.4
langgraph==0.2.59
langchain-community==0.3.13

# Vector Database
pinecone-client==5.0.1

# OpenAI
openai==1.58.1

# Embeddings & Text Processing
tiktoken==0.8.0

# CORS
python-multipart==0.0.18

# Utilities
httpx==0.28.1
